+++ pay_android/android/src/main/kotlin/io/flutter/plugins/pay_android/PayMethodCallHandler.kt
@@ -1,5 +1,5 @@
 /**
- * Copyright 2021 Google LLC
+ * Copyright 2023 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,7 +17,6 @@
 package io.flutter.plugins.pay_android
 
 import android.app.Activity
-import androidx.annotation.NonNull
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.BinaryMessenger
@@ -36,8 +35,9 @@ private const val METHOD_SHOW_PAYMENT_SELECTOR = "showPaymentSelector"
  * A simple class that configures and register a method handler for the `pay` plugin.
  */
 class PayMethodCallHandler private constructor(
-        messenger: BinaryMessenger,
-        activity: Activity,
+    messenger: BinaryMessenger,
+    activity: Activity,
+    private val activityBinding: ActivityPluginBinding?,
 ) : MethodCallHandler {
 
     private val channel: MethodChannel = MethodChannel(messenger, METHOD_CHANNEL_NAME)
@@ -47,31 +47,36 @@ class PayMethodCallHandler private constructor(
         channel.setMethodCallHandler(this)
     }
 
-    constructor(registrar: Registrar) : this(registrar.messenger(), registrar.activity()!!) {
+    constructor(registrar: Registrar) : this(registrar.messenger(), registrar.activity()!!, null) {
         registrar.addActivityResultListener(googlePayHandler)
     }
 
     constructor(
-            flutterBinding: FlutterPlugin.FlutterPluginBinding,
-            activityBinding: ActivityPluginBinding,
-    ) : this(flutterBinding.binaryMessenger, activityBinding.activity) {
+        flutterBinding: FlutterPlugin.FlutterPluginBinding,
+        activityBinding: ActivityPluginBinding,
+    ) : this(flutterBinding.binaryMessenger, activityBinding.activity, activityBinding) {
         activityBinding.addActivityResultListener(googlePayHandler)
     }
 
     /**
      * Clears the handler in the method channel when not needed anymore.
      */
-    fun stopListening() = channel.setMethodCallHandler(null)
+    fun stopListening() {
+        channel.setMethodCallHandler(null)
+        activityBinding?.removeActivityResultListener(googlePayHandler)
+    }
 
     @Suppress("UNCHECKED_CAST")
-    override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
+    override fun onMethodCall(call: MethodCall, result: Result) {
         when (call.method) {
             METHOD_USER_CAN_PAY -> googlePayHandler.isReadyToPay(result, call.arguments()!!)
             METHOD_SHOW_PAYMENT_SELECTOR -> {
                 val arguments = call.arguments<Map<String, Any>>()!!
-                googlePayHandler.loadPaymentData(result,
-                        arguments.getValue("payment_profile") as String,
-                        arguments.getValue("payment_items") as List<Map<String, Any?>>)
+                googlePayHandler.loadPaymentData(
+                    result,
+                    arguments.getValue("payment_profile") as String,
+                    arguments.getValue("payment_items") as List<Map<String, Any?>>
+                )
             }
 
             else -> result.notImplemented()