+++ pay/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 1.1.2 (2023-07-31)
+* Update `pay_android` to 1.0.11, which includes lifecycle fixes.
 
 ## 1.1.1 (2023-02-02)
 * Update `pay_android` to 1.0.10, which includes the latest version of the `flutter_svg` package.